Renamed folder structure within the project.
Removed exception classes.
