+++ SkorkEngine/GUI/MainWindow.xaml.cs
@@ -1,6 +1,6 @@
 ﻿using System.Windows;
 
-namespace SkorkEngine.window {
+namespace SkorkEngine.GUI {
     /// <summary>
     /// Interaction logic for MainWindow.xaml
     /// </summary>
