+++ SkorkEngine/packages.config
@@ -1,0 +1,5 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="EntityFramework" version="6.2.0" targetFramework="net472" />
+  <package id="NUnit" version="3.12.0" targetFramework="net472" />
+</packages>
